utils: always get first abstract in get_abstract
The convoluted logic is not needed anymore, as the data now reflect what
the preferred abstract is and it doesn't have to be computed at display
time.
